+++ packages/sane/lib/sane.dart
@@ -1,8 +1,8 @@
 library;
 
 export 'src/exceptions.dart';
-export 'src/sane.dart';
 export 'src/sane_dev.dart';
-export 'src/sane_isolate.dart';
+export 'src/sane_native.dart';
+export 'src/sane_sync.dart';
 export 'src/structures.dart';
 export 'src/utils.dart';
