+++ packages/sane/lib/src/isolate_messages/cancel.dart
@@ -1,5 +1,5 @@
 import 'package:sane/src/isolate_messages/interface.dart';
-import 'package:sane/src/sane.dart';
+import 'package:sane/src/sane_sync.dart';
 import 'package:sane/src/structures.dart';
 
 class CancelMessage implements IsolateMessage {
