+++ packages/sane/lib/src/sane_native.dart
